lehmer(refactor): Use uint64_t instead of int64_t, decouple selecting streams from seeding, and simplify stream state initialzation

+++ include/lehmer.h
@@ -3,7 +3,7 @@
  *
  * @file include/lehmer.h
  *
- * A simple implementation of the lehmer random number generator in pure C
+ * @brief A simple implementation of the Lehmer RNG in pure C
  *
  * Title: Random number generators: good ones are hard to find
  * Paper: https://dl.acm.org/doi/10.1145/63039.63042
@@ -21,7 +21,9 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdlib.h>
+#include <time.h>
 
+// These constants were defined with 32-bits in mind, not 64-bits.
 #define MODULUS    2147483647 // Mersenne prime number used as modulus (2^31 - 1)
 #define MULTIPLIER 48271      // Multiplier for the Lehmer RNG
 #define CHECK      399268537  // Check value for correctness
@@ -29,25 +31,22 @@
 #define A256       22925      // Jump multiplier for stream separation
 #define DEFAULT    123456789  // Default seed value
 
-// if we can only use positive values, why even bother with negative values?
-// doesn't make sense to me. might be better to enforce unsigned values,
-// e.g. use something like uint64_t isntead.
 typedef struct LehmerState {
-    int64_t* seed;        // Current state of each stream
-    size_t   stream;      // Current stream index, defaults to 0
-    size_t   size;        // Dynamic number of streams
-    bool     initialized; // true if RNG is initialized
+    uint64_t* seed;        // Current state of each stream
+    size_t    stream;      // Current stream index
+    size_t    size;        // Number of streams
+    bool      initialized; // True if RNG is initialized
 } lehmer_state_t;
 
-lehmer_state_t* lehmer_create_state(size_t size, int64_t seed);
+lehmer_state_t* lehmer_create_state(size_t size, uint64_t seed);
 void            lehmer_free_state(lehmer_state_t* state);
 
-void    lehmer_set_seed(lehmer_state_t* state, int64_t value);
-int64_t lehmer_get_seed(lehmer_state_t* state);
+void     lehmer_set_seed(lehmer_state_t* state, uint64_t value);
+uint64_t lehmer_get_seed(lehmer_state_t* state);
 
 void lehmer_select_stream(lehmer_state_t* state, size_t stream);
-void lehmer_seed_streams(lehmer_state_t* state, int64_t value);
+void lehmer_seed_streams(lehmer_state_t* state, uint64_t value);
 
-float lehmer_generate(lehmer_state_t* state);
+double lehmer_generate(lehmer_state_t* state);
 
 #endif // LEHMER_H
